+++ scripts/utility/tashaSummon.js
@@ -150,11 +150,38 @@ async function rangedAttack({speaker, actor, token, character, item, args, scope
     await workflow.setAttackRoll(attackRoll);
     queue.remove(workflow.item.uuid);
 }
+function updateSummonInitiative(actor, [combatant]) {
+    for (let c of combatant?.parent?.combatants?.contents.filter(i => i.actorId != actor.id).filter(i => i.actor.flags?.warpgate?.control?.actor === actor?.uuid)) {
+        if (c.initiative === null) {
+            if (game.user.isGM) c.update({'initiative': combatant.initiative - 0.01});
+            else socket.executeAsGM('updateInitiative', c.uuid, combatant.initiative - 0.01);
+        }
+    }
+}
+function updateCompanionInitiative(actor, [combatant]) {
+    let validIds = [];
+    for (let [key, value] of Object.entries(actor.ownership)) {
+        if (key === 'default' && value === 3) return;
+        if (key === 'default') continue;
+        if (value === 3 && game.users.get(key).isGM === false) validIds.push(key);
+    };
+    if (!validIds) return;
+    for (let i of validIds) {
+        for (let c of combatant.parent.combatants.contents.filter(c => c.actorId != actor.id).filter(c => c.actor.ownership[i] === 3)) {
+            if (c.initiative === null) {
+                if (game.user.isGM) c.update({'initiative': combatant.initiative - 0.01});
+                else socket.executeAsGM('updateInitiative', c.uuid, combatant.initiative - 0.01);
+            }
+        }
+    }
+}
 export let tashaSummon = {
     'setupFolder': setupFolder,
     'getCR': getCR,
     'spawn': spawn,
     'createCombatant': createCombatant,
     'meleeAttack': meleeAttack,
-    'rangedAttack': rangedAttack
+    'rangedAttack': rangedAttack,
+    'updateSummonInitiative': updateSummonInitiative,
+    'updateCompanionInitiative': updateCompanionInitiative
 };