+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # 0.10.43 Release:
+## New Content:
+- Settings for auto updating player summoned and owned tokens' initiative under Summons settings
 ## Bug Fixes:
 - Firearm Support
 - Firearm (CR)